Move status functionality out of lib
no longer needed, makes lib/ slimmer
