+++ lib/src/lib.rs
@@ -10,7 +10,6 @@ pub mod reference_values;
 mod kopium;
 #[allow(clippy::all)]
 mod vendor_kopium;
-use k8s_openapi::jiff::Timestamp;
 pub use kopium::approvedimages::*;
 pub use kopium::attestationkeys::*;
 pub use kopium::ingresses as openshift_ingresses;
@@ -26,87 +25,8 @@ pub use vendor_kopium::virtualmachineinstances;
 pub use vendor_kopium::virtualmachines;
 
 use anyhow::{Context, Result, anyhow};
-use conditions::*;
-use k8s_openapi::apimachinery::pkg::apis::meta::v1::{Condition, OwnerReference, Time};
+use k8s_openapi::apimachinery::pkg::apis::meta::v1::OwnerReference;
 use kube::{Api, Client, Resource};
-
-#[macro_export]
-macro_rules! update_status {
-    ($api:ident, $name:expr, $status:expr) => {{
-        let patch = kube::api::Patch::Merge(serde_json::json!({"status": $status}));
-        $api.patch_status($name, &Default::default(), &patch).await
-            .map_err(Into::<anyhow::Error>::into)
-    }}
-}
-
-pub fn condition_status(status: bool) -> String {
-    match status {
-        true => "True".to_string(),
-        false => "False".to_string(),
-    }
-}
-
-pub trait Conditions {
-    fn conditions(&self) -> &Option<Vec<Condition>>;
-}
-
-impl Conditions for TrustedExecutionClusterStatus {
-    fn conditions(&self) -> &Option<Vec<Condition>> {
-        &self.conditions
-    }
-}
-
-impl Conditions for AttestationKeyStatus {
-    fn conditions(&self) -> &Option<Vec<Condition>> {
-        &self.conditions
-    }
-}
-
-impl Conditions for ApprovedImageStatus {
-    fn conditions(&self) -> &Option<Vec<Condition>> {
-        &self.conditions
-    }
-}
-
-pub fn transition_time<S: Conditions>(
-    existing_status: &Option<S>,
-    type_: &str,
-    new_status: &str,
-) -> Time {
-    let get = |s: &S| s.conditions().clone();
-    let conditions = existing_status.as_ref().and_then(get);
-    let find = |c: &Condition| type_ == c.type_ && new_status == c.status;
-    let existing = conditions.and_then(|cs| cs.into_iter().find(find));
-    let time = existing.map(|c| c.last_transition_time);
-    time.unwrap_or(Time(Timestamp::now()))
-}
-
-pub fn committed_condition(
-    reason: &str,
-    generation: Option<i64>,
-    existing_status: &Option<ApprovedImageStatus>,
-) -> Condition {
-    let status = condition_status(reason == COMMITTED_REASON);
-    let type_ = COMMITTED_CONDITION;
-    Condition {
-        type_: type_.to_string(),
-        reason: reason.to_string(),
-        message: match reason {
-            NOT_COMMITTED_REASON_COMPUTING => "Computation is ongoing. Check jobs for progress.",
-            NOT_COMMITTED_REASON_NO_DIGEST => {
-                "Image did not specify a digest. \
-                 Only images with a digest are supported to avoid ambiguity."
-            }
-            NOT_COMMITTED_REASON_PENDING => "Pod is pending, check pods for details",
-            NOT_COMMITTED_REASON_FAILED => "Computation failed, check operator log for details",
-            _ => "",
-        }
-        .to_string(),
-        last_transition_time: transition_time(existing_status, type_, &status),
-        status,
-        observed_generation: generation,
-    }
-}
 
 /// Generate an OwnerReference for any Kubernetes resource
 pub fn generate_owner_reference<T: Resource<DynamicType = ()>>(
